# openapi.yml
openapi: 3.1.1
info:
  title: fairdrop api
  version: 1.1.2
# #1 TODO: вроде как здесь вместо хоста можно использовать переменную. чтобы не хардкодить мой адрес
servers:
  - url: https://plck.ru/fairdrop
paths:
  /folders/create:
    post:
      summary: создать папку
      description: создаёт новую папку и возвращает информацию о ней
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/folder"
        "5XX":
          description: внутренняя ошибка сервера
  /folders/push:
    post:
      summary: аплоад файлов с созданием папки
      description: создаёт новую папку с полученными файлами и возвращает информацию о ней
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/folder"
        '413':
          description: объем данных в форме превышает лимит
        "5XX":
          description: внутренняя ошибка сервера
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
  /folders/{folder}:
    parameters:
      - $ref: '#/components/parameters/pwParam'
    get:
      summary: получить список файлов папки
      description: возвращает список из имен файлов содержащихся в папке
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: папка не найдена
        '5XX':
          description: внутренняя ошибка сервера
    post:
      summary: загрузить файлы в папку
      description: записывает файлы переданные в форме в указанную папку
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: OK
        '404':
          description: запрошенная папка не существует
        '413':
          description: объем данных в форме превышает лимит
        '5XX':
          description: внутренняя ошибка сервера
    delete:
      summary: удалить папку
      description: удаляет папку
      responses:
        '200':
          description: OK
        '404':
          description: запрошенная папка не существует
        '5XX':
          description: внутренняя ошибка сервера
  /folders/{folder}/{filename}:
    parameters:
      - $ref: '#/components/parameters/pwParam'
      - $ref: '#/components/parameters/filenameParam'
    get:
      summary: скачать файл
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: запрошенные файл или папка не найдены

components:
  parameters:
    pwParam:
      description: пароль-идентификатор папки
      name: folder
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/password'
    filenameParam:
      description: имя файла
      name: filename
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/filename'
  schemas:
    password:
      type: string
      minLength: 6
      maxLength: 6
    filename:
      type: string
      minLength: 1
      maxLength: 256
    folder:
      type: object
      required:
        - password
        - creation_date
        - ip
      properties:
        password:
          $ref: '#/components/schemas/password'
        creation_date:
          type: integer
          format: int32
          description: timestamp в формате unix-time в секундах
        ip:
          type: string
          format: ipv4